requirements.txt
arch==7.0.0
einops==0.8.0
gluonts==0.14.3
hurst==0.0.5
joblib==1.4.2
matplotlib==3.8.4
matrixprofile==1.1.10
numba==0.59.1
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2
scipy==1.13.1
statsmodels==0.14.1
tqdm==4.66.4
transformers==4.38.2
tsfresh==0.20.2
torchinfo
pytorch_lightning
lightning